Feat.comment ErrorCode

+++ backend/src/main/java/com/finder/mypet/comment/service/CommentService.java
@@ -5,6 +5,9 @@ import com.finder.mypet.board.domain.repository.BoardRepository;
 import com.finder.mypet.comment.domain.entity.Comment;
 import com.finder.mypet.comment.domain.repository.CommentRepository;
 import com.finder.mypet.comment.dto.request.CommentRequest;
+import com.finder.mypet.common.advice.exception.CustomException;
+import com.finder.mypet.common.response.ResponseCode;
+import com.finder.mypet.review.domain.entity.Review;
 import com.finder.mypet.user.domain.entity.User;
 import com.finder.mypet.user.domain.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
@@ -24,11 +27,8 @@ public class CommentService {
 
     @Transactional
     public void save(String userId, Long boardId, CommentRequest dto) {
-        User user = userRepository.findByUserId(userId)
-                .orElseThrow(() -> new UsernameNotFoundException("존재하지 않는 유저입니다."));
-
-        Board board = boardRepository.findById(boardId)
-                .orElseThrow(() -> new UsernameNotFoundException("존재하지 않는 게시물입니다."));
+        User user = findByUserId(userId);
+        Board board = findByBoardId(boardId);
 
         Comment comment = Comment.builder()
                 .writer(user)
@@ -41,12 +41,12 @@ public class CommentService {
 
     @Transactional
     public void updateComment(String userId, Long commentId, CommentRequest dto) {
-        userRepository.findByUserId(userId)
-                .orElseThrow(() -> new UsernameNotFoundException("존재하지 않는 유저입니다."));
+        User user = findByUserId(userId);
+        Comment comment = findByCommentId(commentId);
 
-        Comment comment = commentRepository.findById(commentId)
-                .orElseThrow(() -> new UsernameNotFoundException("존재하지 않는 댓글입니다."));
-
+        if (!user.getNickname().equals(comment.getWriter().getNickname())) {
+            throw new CustomException(ResponseCode.NOT_AUTHORITY);
+        }
         if (dto.getContent() != null) comment.setContent(dto.getContent());
 
         commentRepository.save(comment);
@@ -54,13 +54,32 @@ public class CommentService {
 
     @Transactional
     public void deleteComment(String userId, Long commentId) {
-        userRepository.findByUserId(userId)
-                .orElseThrow(() -> new UsernameNotFoundException("존재하지 않는 유저입니다."));
+        User user = findByUserId(userId);
+        Comment comment = findByCommentId(commentId);
 
-        commentRepository.findById(commentId)
-                .orElseThrow(() -> new UsernameNotFoundException("존재하지 않는 댓글입니다."));
+        if (!user.getNickname().equals(comment.getWriter().getNickname())) {
+            throw new CustomException(ResponseCode.NOT_AUTHORITY);
+        }
 
         commentRepository.deleteById(commentId);
+    }
+
+    public User findByUserId(String userId) {
+        User user = userRepository.findByUserId(userId)
+                .orElseThrow(() -> new CustomException(ResponseCode.NOT_FOUND_USER));
+        return user;
+    }
+
+    public Comment findByCommentId(Long commentId) {
+        Comment comment = commentRepository.findById(commentId)
+                .orElseThrow(() -> new CustomException(ResponseCode.NOT_FOUND_COMMENT));
+        return comment;
+    }
+
+    public Board findByBoardId(Long boardId) {
+        Board board = boardRepository.findById(boardId)
+                .orElseThrow(() -> new CustomException(ResponseCode.NOT_FOUND_BOARD));
+        return board;
     }
 }
